pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "Email Triage RL environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]
